feat: call 페이지 fsd 디자인 패턴 적용
