+++ src/entities/user/types.ts
@@ -1,0 +1,7 @@
+export type User = {
+  id: number;
+  name: string;
+  status: 'online' | 'away';
+  avatar: string;
+  statusMessage: string;
+};
