+++ src/features/video-call/model/useVideoCall.ts
@@ -1,0 +1,13 @@
+import { useState } from 'react';
+
+export function useVideoCall() {
+  const [isMuted, setIsMuted] = useState(false);
+  const [isVideoOff, setIsVideoOff] = useState(false);
+
+  return {
+    isMuted,
+    setIsMuted,
+    isVideoOff,
+    setIsVideoOff,
+  };
+}
